refactor: change firends and inventory into slices
